refactor(hipsr): remove per-operation headers
Expose operation classes through HipsrOps.h and move LLVM lowering declarations to a conversion header so the public API remains constant as operations grow.

Made-with: Cursor
