+++ include/hip/Dialect/Hipsr/IR/HipsrConstantOp.td
@@ -51,7 +51,7 @@ def Hipsr_ConstantOp : Hipsr_Op<"constant", [Pure]> {
     /// Returns a typed view of the constant data (inline dense / mem source /
     /// file source via the dialect cache). MorphiZen-only contract; works both
     /// pre- and post-externalization. See the CONTRACT comment on the
-    /// definition in HipsrConstantOp.h. Defined out-of-line there.
+    /// definition in HipsrOps.h. Defined out-of-line there.
     template <typename T>
     ::llvm::ArrayRef<T> getDataValues();
   }];
