--- include/hip/Dialect/Hipsr/IR/HipsrMatMulOp.h
@@ -1,11 +1,0 @@
-/*
- * Copyright (C) 2026 Advanced Micro Devices, Inc. All rights reserved.
- * Licensed under the MIT License.
- */
-
-#ifndef HIPSR_MATMUL_OP_H
-#define HIPSR_MATMUL_OP_H
-
-#include "hip/Dialect/Hipsr/IR/HipsrOps.h"
-
-#endif // HIPSR_MATMUL_OP_H
