+++ lib/Dialect/Hipsr/IR/HipsrAddOp.cpp
@@ -3,7 +3,8 @@
  * Licensed under the MIT License.
  */
 
-#include "hip/Dialect/Hipsr/IR/HipsrAddOp.h"
+#include "hip/Conversion/HipsrToLLVM/HipsrToLLVM.h"
+#include "hip/Dialect/Hipsr/IR/HipsrOps.h"
 
 #include "hip/Dialect/Hipsr/IR/HipsrLLVMLoweringUtils.h"
 #include "hip/Dialect/Hipsr/IR/HipsrShapeRegionInterface.h"
