+++ lib/Dialect/Hipsr/IR/HipsrEmptyOp.cpp
@@ -3,9 +3,7 @@
  * Licensed under the MIT License.
  */
 
-#include "hip/Dialect/Hipsr/IR/HipsrEmptyOp.h"
-
-#include "hip/Dialect/Hipsr/IR/HipsrEmptyYieldOp.h"
+#include "hip/Dialect/Hipsr/IR/HipsrOps.h"
 
 #include "mlir/Dialect/Tensor/IR/Tensor.h" // tensor::EmptyOp
 #include "llvm/ADT/STLExtras.h"
