+++ lib/Dialect/Hipsr/IR/HipsrPlaceholderOp.cpp
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-#include "hip/Dialect/Hipsr/IR/HipsrPlaceholderOp.h"
+#include "hip/Dialect/Hipsr/IR/HipsrOps.h"
 
 using namespace mlir;
 using namespace mlir::hipsr;
