+++ lib/Dialect/Hipsr/IR/HipsrPoolDomainOp.cpp
@@ -3,11 +3,7 @@
  * Licensed under the MIT License.
  */
 
-#include "hip/Dialect/Hipsr/IR/HipsrPoolDomainOp.h"
-
-// Generated SingleBlockImplicitTerminator methods need the complete
-// PoolDomainYieldOp class from HipsrPoolDomainYieldOp.h.
-#include "hip/Dialect/Hipsr/IR/HipsrPoolDomainYieldOp.h"
+#include "hip/Dialect/Hipsr/IR/HipsrOps.h"
 
 #include "llvm/Support/ErrorHandling.h"
